+++ src/Symfony/Component/Mailer/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+6.4
+---
+
+* Add file DSN used as `file://path/to/relative/file.txt` or `file:///path/to/abolute/file.txt`
+
 6.3
 ---